requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-extensions==4.1
python-decouple==3.8
psycopg2-binary==2.9.11
whitenoise==6.12.0
gunicorn==25.1.0
pydotplus==2.0.2
pytest==8.4.2
pytest-django==4.11.1